support for loading schema and dataset based on CQL - removed dependencies from hector in CQL related classes
