+++ src/main/java/org/cassandraunit/CassandraUnit.java
@@ -4,18 +4,14 @@ import me.prettyprint.hector.api.Cluster;
 import me.prettyprint.hector.api.Keyspace;
 import me.prettyprint.hector.api.factory.HFactory;
 import org.cassandraunit.dataset.DataSet;
-import org.cassandraunit.utils.EmbeddedCassandraServerHelper;
-import org.junit.rules.ExternalResource;
 
-public class CassandraUnit extends ExternalResource {
+public class CassandraUnit extends BaseCassandraUnit  {
     public Cluster cluster;
     public Keyspace keyspace;
-
     private DataSet dataSet;
 
     public static String clusterName = "TestCluster";
     public static String host = "localhost:9171";
-    protected String configurationFileName;
 
     public CassandraUnit(DataSet dataSet) {
         this.dataSet = dataSet;
@@ -30,20 +26,7 @@ public class CassandraUnit extends ExternalResource {
         this.host = host;
     }
 
-
-	@Override
-    protected void before() throws Exception {
-        /* start an embedded Cassandra */
-        if (configurationFileName != null) {
-            EmbeddedCassandraServerHelper.startEmbeddedCassandra(configurationFileName);
-        } else {
-            EmbeddedCassandraServerHelper.startEmbeddedCassandra();
-        }
-
-        /* create structure and load data */
-        load();
-    }
-
+    @Override
     protected void load() {
         DataLoader dataLoader = new DataLoader(clusterName, host);
         dataLoader.load(dataSet);
